// app/src/renderer/api/gen/truss/v1/resources_pb.ts
export class GVR extends Message<GVR> {
  /**
   * @generated from field: string group = 1;
   */
  group = "";

  /**
   * @generated from field: string version = 2;
   */
  version = "";

  /**
   * plural
   *
   * @generated from field: string resource = 3;
   */
  resource = "";

  constructor(data?: PartialMessage<GVR>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "truss.v1.GVR";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "group", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "version", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "resource", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): GVR {
    return new GVR().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): GVR {
    return new GVR().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): GVR {
    return new GVR().fromJsonString(jsonString, options);
  }

  static equals(a: GVR | PlainMessage<GVR> | undefined, b: GVR | PlainMessage<GVR> | undefined): boolean {
    return proto3.util.equals(GVR, a, b);
  }
}